+++ src/main/java/backend/general/Factory.java
@@ -1,6 +1,6 @@
 package backend.general;
 
-import backend.general.abstractTools.Config;
+import backend.general.config.Config;
 
 /**
  * Created by Felix on 08.10.2017.
@@ -8,6 +8,6 @@ import backend.general.abstractTools.Config;
 public class Factory {
 
     public static Config getConfig(){
-        return backend.general.abstractTools.Config.getInstance();
+        return backend.general.config.Config.getInstance();
     }
 }
