+++ src/main/java/backend/general/config/Config.java
@@ -1,4 +1,4 @@
-package backend.general.abstractTools;
+package backend.general.config;
 
 /**
  * Created by Felix on 08.10.2017.
@@ -7,7 +7,7 @@ public abstract class Config {
     private static Config instance;
 
     public static Config getInstance(){
-        if (instance==null) instance = new backend.general.tools.ConfigBuilder().build();
+        if (instance==null) instance = new backend.general.config.impl.ConfigBuilder().build();
         return instance;
     }
