+++ src/main/java/backend/general/config/ConfigBuilder.java
@@ -1,4 +1,4 @@
-package backend.general.abstractTools;
+package backend.general.config;
 
 /**
  * Created by Felix on 09.10.2017.
