+++ src/main/java/backend/general/config/impl/ConfigBuilder.java
@@ -1,7 +1,7 @@
-package backend.general.tools;
+package backend.general.config.impl;
 
 
-import backend.general.abstractTools.Config;
+import backend.general.config.Config;
 import com.google.gson.Gson;
 
 import java.io.FileNotFoundException;
@@ -10,7 +10,7 @@ import java.io.FileReader;
 /**
  * Created by Felix on 09.10.2017.
  */
-public class ConfigBuilder implements backend.general.abstractTools.ConfigBuilder {
+public class ConfigBuilder implements backend.general.config.ConfigBuilder {
 
     String CONFIG_PATH= "conf/controller.json";
 
@@ -19,7 +19,7 @@ public class ConfigBuilder implements backend.general.abstractTools.ConfigBuilde
         try {
             FileReader reader = new FileReader(CONFIG_PATH);
             Gson gson = new Gson();
-            return gson.fromJson(reader, backend.general.tools.Config.class);
+            return gson.fromJson(reader, backend.general.config.impl.Config.class);
         } catch (FileNotFoundException e) {
             e.printStackTrace(); //ToDo handle exception
         }
